* master: (345 commits)
  Update dev notes for the coming release
  Increase patchlevel to 3.9.4
  doc fixes
  Added some AIJSELL tests that mirror the SELL tests.
  Bib: More refs
  Removed use of hardwired -1 to represent PETSC_NULL_XXX object, since the null value is now 0.
  fix memleaks flagged by valgrind
  ex62: disable long running test and workaround differences in iteration counts on various builds.
  Add FAQ we can point to when users get error messages such as: "PETSc was configured with one OpenMPI mpi.h version but now appears to be compiling using a different OpenMPI mpi.h version"
  swith test from mpich to openmpi as mpich ch3:sock is unuserable per PetscShmCommGet()
  DM: Forgot to move include statement
  fix nightlybuild issues - disable hypre [for openmp] and fix uninitialized variables
  tests: enable openmp and hwloc with one of the mumps tests
  Default BoomerAMG smoothers
  Configure: Revert changes from debb90c, so that we are back to the original Libmesh
  Configure: Reverted change for Libmesh from 83cd5d7
  Split vecimpl.h into vecimpl.h and vecscatterimpl.h
  SNES ex62: Turned off some tests for other precisions and shortened output of sensitive tests
  Use mmap to allocate the shared pthread barrier if available
  Add configure test for functional mmap
  ...

Conflicts:
	include/petscconvest.h
	src/snes/examples/tutorials/ex13.c
	src/snes/interface/snes.c
	src/snes/utils/convest.c

+++ src/sys/classes/draw/impls/tikz/tikz.c
@@ -125,7 +125,7 @@ static PetscErrorCode PetscDrawEllipse_TikZ(PetscDraw draw,PetscReal x,PetscReal
   win->written = PETSC_TRUE;
   rx = a/2*(draw->port_xr-draw->port_xl)/(draw->coor_xr-draw->coor_xl);
   ry = b/2*(draw->port_yr-draw->port_yl)/(draw->coor_yr-draw->coor_yl);
-  ierr = PetscFPrintf(PetscObjectComm((PetscObject)draw),win->fd,"\\fill [color=%s] (%g,%g) circle [x radius=%g,y radius=%g];\n",TikZColorMap(c),XTRANS(draw,x),YTRANS(draw,y),rx,ry);CHKERRQ(ierr);
+  ierr = PetscFPrintf(PetscObjectComm((PetscObject)draw),win->fd,"\\fill [color=%s] (%g,%g) circle [x radius=%g,y radius=%g];\n",TikZColorMap(c),XTRANS(draw,x),YTRANS(draw,y),(double)rx,(double)ry);CHKERRQ(ierr);
   PetscFunctionReturn(0);
 }
